[X86] Use correct register class for AVX512VL in selectBroadcastFromGPR. Use scalar register class and instructions for scalar GPR->FP copies.
This will enable registers 16-31 to be used for the broadcst.

The scalar register class stuff probably doesn't have much of an
effect. Maybe spill size calculation if this gets spilled. It
does match what isel would generate and ensures we have a scalar
register class coming out to match the instruction that uses it.

Add MIR test to show the register class and instruction names.

Change-Id: Ib243ef08173bb7f5d5b0c3136168b95fbfd13260
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/251514
Reviewed-by: Topper, Craig <craig.topper@intel.com>
Tested-by: Topper, Craig <craig.topper@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>