[VPlan][LoopCFU] Don't skip VPExternalUser live-outs
Technically, that can't be encountered because we create VPExternalUser's in
loop-vectorization pipeline only (so far), and in such pipeline we require
incoming LLVM IR to be in LCSSA form. However, the logic implemented in the
transformation is clearer with this change and it would also be the correct
thing if LCSSA form wasn't enforced.

Change-Id: I898005e6cf195c70e08d727d1fa0d8142a75be4f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/251570
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>