[X86] Avoid a linear scan through a lookup table recognizing instructions in the global FMA pass.
This patch makes use of some of regularities in the encodings
of the instructions we care about in order to quickly identify
them using the data in TSFlags.

I'm not very happy with the code that ignores _Int instructions
for scalar MUL/ADD/SUB. For FMA we had the information statically
coded into the FMA tables we use for commuting. But for the others
I resorted to checking the register class in the instruction
description.

Note the X86InstrFMA3 logic does a quick check using TSFlags to
determine if its an FMA instruction. And then does a binary search
to locate the FMA3Group the instruction belongs to.

I've restored the EVEX/VEX split in the FMA table
and have divided each table into 15 subtables of
6 or 8 VT entries. Linear scans are done through the
subtables. Picking which of the 15 tables is direct.

Change-Id: If867a68c6c70df3c4803c7593826508312e85b22
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/251977
Reviewed-by: Topper, Craig <craig.topper@intel.com>
Tested-by: Topper, Craig <craig.topper@intel.com>