[Driver] Link in Intel libraries for DPC++
Addresses CMPLRLLVM-20168 - We recently added -fveclib=SVML to DPC++
compilations.  This requires that the SVML libraries are also pulled in.
due to the dependence the SVML libs have on libirc, go ahead and just
pull in all of the Intel libs with DPC++ and also provide the proper
lib location.

Change-Id: I54d391139bf08ed41c48aa71e479a69d315a6a50
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/251580
Reviewed-by: Toguchi, Michael D <michael.d.toguchi@intel.com>
Tested-by: Toguchi, Michael D <michael.d.toguchi@intel.com>