[NFC][VPlan] Relax the restriction that VPCallInstruction should always be master.
VPlan's HIR decomposer recently added a restriction that all calls
from incoming HIR are expected to be standalone HLInsts thereby
enforcing all VPCalls to be master VPInstructions. However this
is not always true since calls of the following format can be
generated by HIR transforms -

(%TempArray)[i1] = call(%arg1, %arg2,...)

This patch relaxes the restriction in decomposer by allowing the
VPCall to be a slave VPInstruction of the generated store. Even for
this case, the underlying HLDDNode of the VPCall will still be a
HLInst which has the corresponding llvm::CallInst as underlying
LLVM instruction. This change is still in compliance with the other
major assumption in VPlan - Every VPCall has an underlying
llvm::CallInst which can be accessed for more information about the
call.

Change-Id: I3e06a44a992a411251f3f26b5c99bef68998ceb2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/251951
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>