[VPlan] Fix HIR VPValue based code generation of uniform load
Designated reviewers -- Primary: Karthik, Secondary: Matt

Changes here implement code generation of load where the pointer operand being
loaded from is uniform. We do a scalar load using the pointer operand and the
value is broadcast to generate the vector value. If the load is under a mask,
the load is done conditionally only if mask is non-zero.

Change-Id: I9808be4cafae21904ff1eb70bb6ea8a145e21403
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/252049
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>