[DPCPP][Host] Port over BarrierInFunction from OpenCL
This is a part of effort to enable DPCPP kernels on the HOST through
WGLoopCreator/Barrier approach.

The goal is to enable a subset of benchmarks by doing this, not provide
product quality from day #1.
This port does not aim to port complete functionality from OpenCL to
DPCPP, only core functionality that is needed achieve the above.
So unless problems detected in review are critical to this goal, I'd
rather file JIRAs and address them later.

The pass is a prerequisite of main Barrier pass.

Change-Id: I472f233b822fe2707c761890d33e1347e33da13b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/252101
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>