[OCLVecClone] Suppress shl/bin_op.../ashr pattern for TID calls
Users may usually write code such as

  int id = get_*_id();
  ... = array[id+1];

and 'id' can be first truncated to i32, then extended to i64 and added
with 1. After InstCombine pass, 'id+1' becomes

  %shl = shl %id, 32
  %add = add i64 %shl, (1<<32)
  %id.plus.one = ashr %add, 32

In this patch, when -less-than-two-gig-max-work-group-size and/or
-less-than-two-gig-max-global-work-size are specified, we try to detect
TID use patterns with arbitrary number of binary operations surrounded
by shl and ashr, where binary operations can be add, sub and mul. And
then we remove the shl and ashr, and shift back constants in binary
operations. This will benefit later vectorization.

Change-Id: I0e5242b32de53007aa14a3a51f54112dc06f7375
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/252813
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>