[DTRANS] CMPLRLLVM-5594: Fix for analyzing structures with  padding  (Part 1) [XMAIN]
This change set identifies when the CFE adds padded structures for
base classes due to ABI. For example, consider the following classes:
 class A {
   int a;
   int b;
 }

 class B : public A {
   int c;
 }

Also consider that there is an instantiation of A and B, then we will
see something like this in the IR:

 %class.A = type <{i32, i32, [ 4 x i8 ]}>
 %class.A.base = type <{i32, i32}>
 %class.B = type {%class.A.base, i32}

This change set identifies when we have these base and padded types and
maps them. The safety data found for one type will also be enabled in
the related type.

Change-Id: I3ec88c77bfeb76ed1df870a461d9e0e9f94b0735
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/253245
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>