[VPlan] Fix for all-zero bypass region collection for phis.
This patch terminates block collection for a region when predecessors
of a block are not found in the region. Blocks for the region are now
collected on-the-fly so that the predecessors can be checked. In the
case of loops, all blocks are added all at once to avoid having to deal
with backedges. Also split the region collection for loops and non-loops
into separate functions that are controlled from two different switches.

This patch also sets up the ability to fix an issue where region collection
tries to include blocks containing CondBits with two forward edges. See
https://git-amr-2.devtools.intel.com/gerrit/#/c/248816/.

Change-Id: Id8c10338d3d042fa9f8ce8d732bebcbf355491a7
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/252666
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>