[VPlan] Fix all-zero bypass to exclude blocks from region.
This patch fixes a problem where a block with two forward edges is
inserted into the region. This kind of thing happens when we have
a block containing only a CondBit. Including these blocks without an
additional check to see if the post-dominator has a block-predicate
under the influence of the region can result in forming incorrect
regions. For now, stop region collection at these types of blocks.

Change-Id: I865287f4a52bbac08a0de48bc1bb26fba0c9b760
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/252708
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Masten, Matt <matt.masten@intel.com>
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>