[call-tree-cloning] Delete old call instructions after specializing call site
This change fixes segfault in LTO plugin on denbench/aes@opt_base_xE with the
following patch applied https://git-amr-2.devtools.intel.com/gerrit/#/c/253265/
Segfault is happening on the following code from LLVMContextImpl destructor
which tries to dump leaked instructions with metadata

  // Check for metadata references from leaked Instructions.
  for (auto &Pair : InstructionMetadata)
    Pair.first->dump();
  assert(InstructionMetadata.empty() &&
         "Instructions with metadata have been leaked");

Call instruction with no parent left by call-tree-cloning pass was the one
which triggered the segfault.

Creating a small reproducer for this problem would be quite difficult, so I am
asking for an exception to commit this change with no LIT test.

Change-Id: I6decec0f3b7a0ba63d3e809ef0a20d6663e88d6b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/253608
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Bobrovsky, Konstantin S <konstantin.s.bobrovsky@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>