[VPlan] Improvement to folding of IV convert operations into a canon expression
The IV in a canon expression can be of a type that is different from the
canon expression's source type. A canon expression for representing the value
zext(trunc(IV1, i3), i64) looks like [IV1], where the src type of the CE is
set to i3 and the dest type is set to i64. In order to preserve linear canon
expression in generated vector code for such cases, we flag such IV converts
during decomposition to allow us to fold both converts into the canon expression
preserving linearity.

Change-Id: I62c085f7c798698d5bbbbf96f5e1e17bc3db3f08
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/253606
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>