CMPLRLLVM-20261: Introduce llvm.for.cpystr intrinsic
For some reasons, Fortran frontend can't emit llvm.memmove and llvm.memset
directly to copy constant string with padding if dstlen and srclen are different
size. It always calls for_cpystr to do this. We can turn this function to an
intrinsic function so that it can be optimized to llvm.memmove and llvm.memset
intrinsic.

Change-Id: I4ea7877b8d0aebaf65d30568a6ae956e65d8cccb
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/254107
Reviewed-by: Xiao, Wei3 <wei3.xiao@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>
Tested-by: Haohai Wen <haohai.wen@intel.com>