CMPLRLLVM-20234: do not use file-system's file ids for OpenMP mangling.
This is needed to avoid build-to-build variations for OpenMP offload
performance testing.  PTA team will use this mode for running CS
analysis using less machine resources.

I am not sure how to write a LIT test for this, since I need to
compile two files with the same name in different directories.
LIT seems to compile files with absolute names, so the hash
will always be different for the two files.

I will add a TC test after pushing this.

Change-Id: I66e0859244d4713b06fc0542b1ba1322c9cd017b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/254214
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Tested-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>