[DTrans] CMPLRLLVM-5594: Fix for analyzing structures with padding (Part 2)
This is the second change set for analyzing the structures with ABI padding.
This change set is to invalidate the relationship between base and padded
structures if there is an illegal access to the padded field. We can catch
these illegal access by checking the FieldInfo if the field was read, written,
address taken, has ccomplex use, or mismatched element access.

Change-Id: I5f7e3afa9536281820e2008ede977c0f16e86f01
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/254534
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>