[VPlan] Generate wide load/store for -1 stride with appropriate vector reversal
Changes in this patch implement generation of wide loads and stores for -1
stride case. The value loaded or to be stored is reversed appropriately
after the load or before the store. The pointer operand of the load/store
is adjusted accordingly. If the load/store is under a mask, the mask
value is reversed as well.

Change-Id: Ifa205f10a6f84177f5959b57ce7eaf1c81729c6f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/254808
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>