DTRANS CMPLRLLVM-5594: Mark as bad casting for related types when moving data from a source to a destination with the same type
This change set is for marking as "Bad casting for related types" in
DTrans when a data is moved between two pointers of the same type.
For example:

  %1 = getelementptr inbounds %struct.test, %struct.test* %VAR,
        i64 0, i32 1
  %2 = bitcast %"inner.struct"* %1 to i64*

  %3 = getelementptr inbounds %struct.test, %struct.test* %VAR2,
        i64 0, i32 1
  %4 = bitcast %"inner.struct"* %3 to i64*
  %5 = load i64, i64* %4, align 8

  store i64 %5, i64* %2

In the previous example, the %5 is being stored in %2, but the source
and destination are from the same type. This will produce a bad casting,
but is not an issue to prevent merging two related types. Therefore,
this will produce a "bad casting for related types" and any
transformation that depends on bad casting will also check for this
new safety data.

Change-Id: Ic7869a0e8657082681f15ba71a0997240e1191bc
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/254850
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>