Fix attribute preservation of arguments in function call vectorization.
Previously, attributes in function call arguments is directly copied to the
vectorized call. This is not correct because we may add, remove, or reorder
arguments during vectorization. This patch fixes it by tracking attributes of
arguments together with their values and types.

Most of time there is no attribute in arguments of math library call. But for
sincos we usually get "nonnull" in output arguments. In this case copying the
AttributeList directly to the vectorized call would fail the verifier.

JIRA: CMPLRLLVM-19433
Change-Id: Ib721b25f6a36356d99e2f5961e9ffbe309c74d50
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/255636
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Tianqing Wang <tianqing.wang@intel.com>
Reviewed-by: Sotkin, Alexey <alexey.sotkin@intel.com>