[VPlan] Add support for Abs VPInstruction
Changes in this patch add support for generating Abs VPInstruction. This
new instruction is generated for HIR HLInsts that are in isAbs() form. This
helps preserve abs() recognition in generated vector code and reduce
the number of VPInstructions generated which helps recover the small
performance regressions seen in denbench benchmarks with VPValue based
code generation(CMPLRLLVM-20530).

Change-Id: Ic56af1730acad41fe244c2d2f73b050ddf26b743
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/255514
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: Guggilla, Satish <satish.guggilla@intel.com>