[VPO][OpenMP] Handle order clause in loop related directives
Added getters and setters for order clause in loop related directives resulting
from loop construct, which include simd, for, parallel for and distribute
parallel for. Now per OpenMP 5.0 spec, with order(concurrent) is same as
without order clause, and it's regarded as a no-op, so the clause is parsed but
ignored.
Added order clause support in print routine for related WRNs.
Fixed 2 typos.

JIRA: CMPLRLLVM-20546

Change-Id: Id65c93108b5484d6b0021f471b12558d7ad82f39
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/255703
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>