[DTRANS] CMPLRLLVM-5594: Relax the conditions for collecting the zero element in a inttoptr instruction
The conditions to collect the zero element from a intotoptr instruction is
a little bit conservative. It won't be able to catch this example:

%class.OuterClass = type {%class.TestClass, %"class.std::vector.12"}
%"class.std::vector.12" = type { %"struct.std::_Vector_base.13" }
%"struct.std::_Vector_base.13" = type { %"struct.std::_Vector_base<TestClass, std::allocator<TestClass>>::_Vector_impl.67" }
%"struct.std::_Vector_base<TestClass, std::allocator<TestClass>>::_Vector_impl.67" = type { %class.OuterClass*, %class.OuterClass*, %class.OuterClass* }
%class.TestClass = type {i64, i64, i64}

define internal %class.OuterClass* @loadData(%class.OuterClass* nocapture %arg) {
bb:
  %tmp = getelementptr inbounds %class.OuterClass, %class.OuterClass* %arg, i64 0, i32 1
  %tmp2 = bitcast %"class.std::vector.12"* %tmp to i64*
  %tmp3 = load i64, i64* %tmp2
  %tmp4 = inttoptr i64 %tmp3 to %class.OuterClass*
  ret %class.OuterClass* %tmp4
}

The GEP in %tmp access field 1 in %class.OuterClass and then it is used
in a combination of bitcast, load and inttoptr instructions to get
the 0 element from %"class.std::vector.12".

The analysis for the inttoptr instructions (%tmp4) is conservative and
only consider the 0 element of the parent structure (%class.OuterClass)
and not the field that it is being actually lodaded (%"class.std::vector.12").
This change set is for relaxing these conditions to collect the information
needed from the actual field rather the parten structure.

Change-Id: I05bfab4060cbb4c30d92c8be900f1d42b1f19264
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/255797
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>