[NFC][Paropt][OpenMP][Target] Add an option to disable simulated support of omp_get_num_threads in target constructs.
The flag "-vpo-paropt-simulate-get-num-threads-in-target=false" can be used
to disable the generation of calls to kmpc_spmd_push/pop_num_threads, which
are used to simulate support for omp_get_num_threads in target constructs.

This might be preferable in some cases as emitting those calls can cause a
performance impact.

Change-Id: I6fd6622307535638f0ceff37106b8c17ee39597e
JIRA: CMPLRLLVM-20123
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/256264
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>