[DTRANS] Add mismatched element access and unsafe pointer store for related types when loading and/or the zero element in nested structures
This change set adds two new safety bits, but they are related to
each other. Mismatched element access for related types basically
means that an access to a zero element by bitcast and load,
or storing into a zero element with the same type won't affect
the analysis for related types. For example:

   %"class.outer" = type { %"class.inner1" }
   %"class.inner1" = type { %"class.inner2", %"classinner2"}
   %"class.inner2" = type { %class.TestClass*, %class.TestClass*}
   %class.TestClass = type { i64, i64, i64}

   %1 = getelementptr inbounds %"class.outer", %"class.outer"* %VAR,
            i64 0, i32 0
   %2 = bitcast %"class.inner1"* %1 to i64*
   %3 = load i64, i64* %2

   %4 = getelementptr inbounds %"class.outer", %"class.outer"* %VAR2,
            i64 0, i32 0
   %5 = bitcast %"class.inner1"* %4 to i64*
   store i64 %3, i64*%5

In this case the load instruction in %3 and the store instruction
access zero elements, therefore this is not a mismatched element
access.

Also, unsafe pointer store for related types basically means that
the store instruction is copying data into the zero element and
this data is the same type, 0 or null.

Change-Id: I526870efdf4b42732cbd8f1558dfb8c116fa313e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/256693
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>