[Paropt][OpenMP][Target] For use_device_ptr operands which also have a map clause, don't create an additional map.
An implicit map with type TGT_MAP_RETURN_PARAM is usually created for
use_device_ptr clause operands. However, it is not needed when
there is a map sent in from the frontend. In which case, the
existing MapType flag of the map clause is updated to also
include TGT_MAP_RETURN_PARAM.

This is needed for Fortran, when the user maps as well as marks
a dope vector with use_device_ptr clause:

  integer, allocatable :: x(:)
  ...
  !$omp target data map(x) use_device_ptr(x)

The same code will be triggered for use_device_addr in C/C++,
as in Paropt, use_device_addr is equivalent to use_device_ptr
(without a PTR_TO_PTR modifier). For example:

  int x;
  ...
  #pragma omp target data map(x) use_device_addr(x)

Change-Id: I4fc8fd9aeae57cf8ae48c059245c3dcd0f7795a7
JIRA: CMPLRLLVM-20798
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/256351
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>