Remove pipelining between lld-link and FileCheck in svml test cases
This change set removes the pipelining between lld-link and
FileCheck. It seems that Windows is processing the next instruction
without releasing the previous process and cause a missing
stdout/stderr exit error. This is a flaky problem.

Also added a check that prevents running the 64bits version of
the test case into a 32bits environment.

Change-Id: I34f433733a8d54438a62481e3a1acb3168958907
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/256775
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>