[DTRANS] CMPLRLLVM-5594: Prevent bad alloc size if the size for new is a multiple of the type's size
This change set is for preventing bad alloc size if we know that the
input size for a call to new is a multiple of the type size. For example:

   %tmp9 = sub i64 %tmp8, %tmp2
   %tmp10 = sdiv exact i64 %tmp9, 48

   %tmp13 = invoke noalias nonnull i8* @_Znwm(i64 %tmp9)

   %tmp14 = bitcast i8* %tmp13 to %class.TestClass*

Assume that the size of %class.TestClass is 48 bytes. From the bitcast
we know that the "new" instruction is used to allocate a new
%class.Testclass object. The size for the allocation is given by the
subtraction in %tmp9. This subtraction is guarded with an exact division
(%tmp10), which means that the result of the subtraction will be a
multiple of 48.

Change-Id: Ie7fdab8be1adadd6988cf765288cfa319ab89937
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/256753
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>