Split CPU RT and FPGA emulator on building step
With this patch within a single build process 2 sets of libraries
are being built: CPU RT and FPGA emulator. On installation step they
are being deployed into to different folders: lib/x64 and lib/emu.
cl.cfg file is being re-generated for the emulator platform.

That is a naive approach and it's better to merge all RT libraries into
one with changing name of the output library on the installation step
(icd loads OpenCL RT entry point by _dynamic loading_ mechanism, which
doesn't care about the actual SONAME of the library in the appropriate
ELF file).

Change-Id: Ieebb79ef7ae10cc180ddbe5ab6ab135dff9f789a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/244914
Tested-by: Dai, Chunyang <chunyang.dai@intel.com>
Reviewed-by: Wenju He <wenju.he@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>