[SYCL] Deny support of SPV_INTEL_usm_storage_classes for all targets but FPGA
Disable SPV_INTEL_usm_storage_classes by default since it adds new
storage classes that represent global_device and global_host address
spaces, which are not supported for all targets. With the extension
disabled the storage classes will be lowered to CrossWorkgroup storage
class that is mapped to just global address space.

The extension is enabled only for FPGA with a flag 'enable-usm-address-spaces'
passed to the driver.

Change-Id: Ia3c864a8761b0232b6f41fa6be7de435ebee1387
Signed-off-by: Dmitry Sidorov <dmitry.sidorov@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/256959
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>