CMPLRLLVM-11974 Correct behavior of LibIRC CPUDispatch-
The comunity version of CPUDispatch uses the libgcc/compiler-rt
compatible feature checks, using the same mechanism as
__builtin_cpu_supports. Unfortunately, this list is much smaller than
libirc (the intel alternative).

At one point in the past, we switched CPU-Dispatch to use libirc for
figuring out the type of processor (the same as the
_may_i_use_cpu_feature). However, I apparently inadvertently failed to
stop removing the list of unsupported features from the check-list.

The result is that in a couple of small cases, where a processor only adds a
single feature that isn't checkable via __builtin_cpu_supports,
CPU-dispatch was dispatching to the incorrect variant.

This fixes that by adding a number of new intel-customization blocks to
better keep and sort this information.

Change-Id: Idb07ec876537d882cd24a02ea23b9a2934b5011d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/257033
Reviewed-by: Rao, Premanand M <premanand.m.rao@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>