[Paropt][OpenMP][Target] Fix passing of size of a VLA marked private on parallel inside target.
On the host side, the temporary VAR allocated to capture VLA size at
the parallel construct was being allocated in the beginning of the
parent function, instead of the beginning of the target region.

For device compilation, there needed to be an address-space cast for
the VLA size being captured at the target level.

The IGC support for VLAs is still not present, so this only prevents the
compiler assertions. Behavior at runtime is still unknown.

Change-Id: Ia4244d9deb58fd390c52cfaee3a52a699c0e4fc3
JIRA: CMPLRLLVM-18855
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/257111
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>