[NFC] [CMPLRLLVM-20352]: Improve build time for SCEV's StrengthenNoWrapFlags
StrengthenNoWrapFlags has Intel improvements to prove NAry no-wrap flags
for large (3- and 4-way) sums and products. This works by considering
all possible binary combinations of the sums and proving that the flags
hold in all cases.

Previously a set of multisets of multisets was used to enumerate all
possible combinations, ensuring that we only evaluate the minimum number
of equivalent combinations. With this change, the complete set of
combinations is not generated ahead of time; instead, they are generated
on demand only if a conclusion has not been reached. In the worst case
this can result in testing more combinations, but build time is
generally faster since set/multiset are slow and we typically don't need
to test all combinations to reach a conclusion.

The on-demand partitions are generated in a deterministic order and all
uses of SetVector/std::multiset have been removed, so this change should
maintain deterministic behavior.

Change-Id: If7d8c2f35ba6ca6af652308329d9d0d66185184f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/259602
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>