[VPlan][HIR] Use VPExternalUses to determine live-out temp for reduction finalization.
Currently the live-out temp that needs to be used for lval of
reduction finalization is determined using VPEntities analysis
whose results can be stale before CG. This patch updates VPlan
HIR codegen to use VPExternalUses to determine live-out temp of
reduction. If a reduction is not live-out (possible for
SafeReduction), then finalized value is written to a new temp
since it will not be used anywhere.

Change-Id: I07493e289fdab45cde31a7ecd268c46e0a1e5e03
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/256293
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>