[LSR][NFC] Cache SCEVs for phi nodes to improve compile time
CMPLRLLVM-20685: improves build time for 554.roms_r by ~12%.

This patch improves compile time of LSR pass in the scenario
when the working set of formulas is large and the number of
phi nodes in outer loop is also big.
In such case, LSR will keep on calling getSCEV for phi nodes
of outer loop (see isExistingPhi) which can be cached since
the IR is not changed during the process of filtering
available formulas.

No test provided, since this is an NFC patch. Testing shows
Buildsame on all alloy benchmarks.

Change-Id: Icf458006561869ce947db5e97dedb8b36893abb2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/257373
Reviewed-by: Lin Jr, Chang-Sun <chang-sun.lin.jr@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: Bakhvalov, Denis <denis.bakhvalov@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>