[VPlan][NFC] Remove explicit condition bit cloning.
Current implementation of condition bit cloning is incorrect: we should
not create a new clone of instruction, we should get it from ValueMap
variable.

But since condition bit is represented as a terminator instruction
operand, we don't need to clone it explicitly. This work is done by
VPValueMapper::remapInstruction method.

JIRA: CMPLRLLVM-20923
Change-Id: I89bbe83dfdc75e38dfa4e1ea8d4900542719967f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/257400
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Mitropoulou, Konstantina <konstantina.mitropoulou@intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>