Refactor of #pragma ivdep(get rid of the IVDEP directive form).
This patch generates SYCLIntelFPGAIVDepAttrs from the pragma ivdep
instead of LoopHintAttrs and reuse the SYCL codes in CGLoopHint/CodeGen.

We share the Sema codes of ivdep pragma with the attribute.
As soon as we detect that the pragma ivdep is the HLS type,
we call a different function that processes it as a SYCLIntelFPGAIVDepAttr.

This patch also adds support for the following two JIRA's.

https://jira.devtools.intel.com/browse/CMPLRLLVM-19305:
OpenCL compiler - crash with #pragma ivdep array on while(1) loop

https://jira.devtools.intel.com/browse/CMPLRLLVM-18400:
Support template parameters in ivdep pragmas in C++

Change-Id: I354c3f223e25d5636021806a648dcc5c8610f274
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/257514
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Rice, Michael P <michael.p.rice@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>
Tested-by: Manna, Soumi <soumi.manna@intel.com>