[CLWGLoopCreator] Add the peeling loop
If VOLCANO_LLVM_OPTIONS is set to "-vplan-enable-peeling", vector code
may have a load/store with intel.preferred_alignment metadata. This
patch computes peeling count based on the metadata and create peeling
loop. The load/store will be set alignment which is the same as
preferred_alignment.

This improves DPC++/stream-triad perf by 12.5% on 2-socket CLX with 96
HW threads.

Co-authored-by: Tianyang Zhu <tianyang.zhu@intel.com>

Change-Id: I399398b00ceb48008501c10a158a62c83f651ae3
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/247947
Reviewed-by: wuwei <wei.wu@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Sotkin, Alexey <alexey.sotkin@intel.com>