[CMPLRLLVM-20229] Temporarily disable HIRConditionalLoadStoreMotion by default
HIRConditionalLoadStoreMotion was added as a precursor for
CMPLRLLVM-11706 in order to improve LBM performance, but on its own this
pass leads to a small regression in LBM performance due to
inconsistencies in the prefetching heuristics and has therefore landed
on the MFL. Because code cut-off is tomorrow and because it seems that
we may not get a waiver for this, I'm changing this pass to disabled by
default as a temporary stopgap until the rest of the LBM work is
finished.

This pass should be re-enabled as part of the LBM work when it is ready.

Change-Id: Ia8cfeecd2baef5d99bfc14a84397557a154ac295
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/258353
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>