[1/3][VPlan][NFC] Adding VPLiveInValue and VPLiveOutValue.
VPLiveInValue is a wrapper for VPLoopEntities incoming values. It is used
to link incoming values from merge blocks and to defend original incoming
values from unprofessional use.
VPLiveOutValue is a wrapper for VPExternalUse in VPlan. It is used to link
outgoing values with VPExternalUse w/o adding outgoing values as operands
of VPExternalUse.
Added containers for VPLiveOutValues and VPLiveInValues in VPlan.
Added container for original incoming values in VPExternalValues.

Change-Id: I72b4f9085d9029b5fbc7a2bfd7a1b09e59bd7565
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/260609
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>