[HIRParVecAnalysis] Use HLNodeVisitor instead of HLInnerToOuterLoopVisitor
HIRParVecAnalysis is the only user of HLInnerToOuterLoopVisitor,
nobody maintains this class. Due to its bugs
ParVecVisitor::visit(HLInst*) was never visited.

DDWalk is a better place to hold per-HLInst legality checks as it is
taught to ignore preheader/postexit blocks, so such checks are
consolidated there for consistency and correcteness.

Also fixing out-of-boundary access to diagnostics array that was hidden
behind the above problem.

Change-Id: Ie57cd8fbab0cf67f90dd0c8882ec7d77f6bfbc1e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/259603
Tested-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Sotkin, Alexey <alexey.sotkin@intel.com>