[VPlan] Correct isLoopHeaderPHI for non-loop phis
We don't have non-loop phis in our loop vectorization pipeline so the change is
untestable. I could have made it an assert to just make Klocwork happy and
declare it NFC, but this change seems to be better and will just work if we will
ever implement SVA/CG for function vectorization.

Change-Id: Ie6f5b14b009204f38934509fe57cc294f4aef33e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/259304
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>