[SCEV]: Perform constant folding before StrengthenNoWrapFlags for sums
This change resulted from CMPLRLLVM-20352, but is not its primary
resolution.

Attempt any constant folding of sums before attempting to strengthen
wrapping flags. (Rather than after.) This can help the build time of
StrengthenNoWrapFlags by allowing it to reason about a smaller set of
operands.

It can also help the precision of StrengthenNoWrapFlags in two ways:

1. Constant folding may bring the total number of operands below the
heuristic threshold for disabling various reasoning within
StrengthenNoWrapFlags.

2. Heuristics aside, constant folding the SCEV may eliminate a genuinely
wrapping binary combination, allowing the effective expression to not
wrap. For example, (1024 + -1023 + %n) might contain a wrapping binary
sum if implemented as ((1024 + %n) + -1023), but not if simplified to (1
+ %n). In this case, no-wrap can be understood only after constant
folding.

Products undergo a similar strengthening/folding sequence, but this
patch doesn't make the corresponding change there.

Change-Id: I0167ba28ef4b0fa3bff7a3a6e7d2e88312894063
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/259651
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>
Tested-by: Creech, Timothy M <timothy.m.creech@intel.com>