CMPLRLLVM-18457: Allow local_mem_size to include decayed pointers
For some unknown reason, local_mem_size and buffer_location (implemented
by the OpenCL team, but nooone seems to remember why) use
ParmVarDecl->getOriginalType, to do diagnosis, which removes decayed
parameters.

In reality, we should just be using the type of the parameter after
decay, so this patch changes it to just getType. Array->pointer decay is
the only case that should apply here, so the effect of this is simply to
allow array types for these params.

Change-Id: I4161dd2a44a94d6bb6e7a6711acf0f802631b1cd
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/259783
Reviewed-by: Rao, Premanand M <premanand.m.rao@intel.com>
Reviewed-by: Manna, Soumi <soumi.manna@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>