[VPlan] Fix handling of VPInductionInit for pointer-inductions in DA.
This change fixes the stride-computation for pointer-induction for both
privates and non-privates. The strides were computed incorrectly when
the patch - https://git-amr-2.devtools.intel.com/gerrit/#/c/252530/
was commited.

Other change in the patch is to rename 'InductionBinop' to just generic
'InductionOp' to better reflect the nature of inductions in case of
pointer-inductions.

Change-Id: I6a2f5b5441fd3e541c90e0fa62d176bb7f3644c9
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/259650
Tested-by: Pushkar Ratnalikar <pushkar.v.ratnalikar@intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>